login page post method and minor changes
