+++ src/web/static/css/index.css
@@ -1,0 +1,90 @@
+* {
+    margin: 0;
+    padding: 0;
+    box-sizing: border-box;
+}
+
+body {
+    font-family: "Audiowide", sans-serif;
+    background-color: #f5f5f5;
+    color: #333;
+    text-align: center;
+}
+
+.audiowide-regular {
+    font-family: "Audiowide", serif;
+    font-weight: 400;
+    font-style: normal;
+  }
+
+  header {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    background-color: #edc885;
+    color: white;
+    padding: 1em 2em;
+}
+
+.login-button {
+    width: 40px;
+    height: 40px;
+    cursor: pointer;
+    transition: transform 0.3s ease;
+}
+
+.login-button:hover {
+    transform: scale(1.1);
+}
+main {
+    margin: 2em auto;
+    width: 60%;
+    background-color: white;
+    padding: 2em;
+    box-shadow: 0 0 10px rgba(0,0,0,0.1);
+}
+
+form {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+}
+
+label, input, button {
+
+    width: 80%;
+    margin: 0.5em 0;
+}
+
+input {
+    padding: 0.5em;
+    border: 1px solid #ccc;
+    border-radius: 4px;
+}
+
+button {
+    padding: 0.5em;
+    background-color: #cf9e42;
+    color: white;
+    border: none;
+    border-radius: 4px;
+    cursor: pointer;
+}
+
+button:hover {
+    background-color: #cf9e42;
+}
+
+footer {
+    margin-top: 2em;
+    padding: 1em 0;
+    background-color: #cf9e42;
+    color: white;
+    position: fixed;
+    bottom: 0;
+    width: 100%;
+}
+
+form label, form input, form button {
+    opacity: 0;
+}
