+++ src/web/static/css/login.css
@@ -1,0 +1,106 @@
+@import url('https://fonts.googleapis.com/css2?family=Poppins:ital,wght@0,100;0,200;0,300;0,400;0,500;0,600;0,700;0,800;0,900;1,100;1,200;1,300;1,400;1,500;1,600;1,700;1,800;1,900&display=swap');
+
+* {
+    margin: 0;
+    padding: 50;
+    box-sizing: border-box;
+    font-family: 'Poppins', sans-serif;
+    text-decoration: none;
+    list-style: none;
+}
+
+body {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    min-height: 100vh;
+    background: linear-gradient(90deg, #e2e2e2, #c9d6ff);
+}
+
+.containerp {
+    font-size: 14.5px;
+    margin: 15px 0;
+}
+
+form{
+    width: 100%;
+}
+
+.form-box{
+    position: absolute;
+    width: 50%;
+    height: 100%;
+    right: 0;
+    background : #fff;
+    display: flex;
+    text-align: center;
+    align-items: center;
+    color: #333;
+    transition: 0.5s ease-in-out 1.2s, visibility 0s 1s;
+    z-index: 1;
+    padding:40px;
+}
+
+.input-box{
+    position: relative;
+    padding: 30px 0;
+}
+
+.input-box input{
+    width: 100%;
+    padding: 13px 50px 13px 20px;
+    border: none;
+    border-radius: 8px;
+    background: #eee;
+    outline: none;
+    color: #333;
+    font-size: 16px;
+    font-weight: 500;
+}
+.input-box ::placeholder-shown{
+    color: #888;
+    font-weight: 400;
+}
+
+.input-box i{
+    position: absolute;
+    right: 20px;
+    top:50%;
+    transform: translateY(-50%);
+    font-size: 20px;
+}
+
+.forgot-link{
+    margin: 0px 0 0x;
+}
+.forgot-link a{
+    color: #333;
+    font-size: 14.5px;
+}
+
+.btn{
+    width: 100%;
+    height: 50px;
+    background: #333;
+    box-shadow: 0 0 10x rgba(0, 0, 0, 0.1);
+    border-radius: 8px;
+    cursor: pointer;
+    color: #fff;
+    font-size: 16px;
+    font-weight: 600;
+    border: none;
+}
+
+@media screen and (max-width: 650px){
+
+    .form-box{
+        bottom: 0;
+        width: 100%;
+        height: 70%;
+    }
+
+
+    @media screen and (max-width:400px){
+        .form-box{padding:20px;}
+    }
+}
