+++ src/web/templates/index.html
@@ -1,0 +1,44 @@
+<!DOCTYPE html>
+<html lang="it">
+<head>
+    <meta charset="UTF-8">
+    <title>ProMeteo</title>
+    <link rel="preconnect" href="https://fonts.googleapis.com">
+    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
+    <link href="https://fonts.googleapis.com/css2?family=Audiowide&display=swap" rel="stylesheet">
+    <link rel="stylesheet" href="./static/css/index.css">
+    <script src="./static/js/index.js" defer></script>
+</head>
+<body>
+    <header>
+        <a href="/menu">
+            <img src="menu.png" alt="Login" class="login-button">
+        </a>
+        <h1>Benvenuto!</h1>
+        <a href="/login">
+            <img src="login-icon.png" alt="Login" class="login-button">
+        </a>
+    </header>
+    <main>
+        <form>
+            <label for="citta">Città</label>
+            <input type="text" id="citta" name="citta">
+            <br/>
+            <label for="data">data</label>
+            <input type="text" id="data" name="data">
+            <br/>
+            <label for="sezione">sezione</label>
+            <input type="text" id="sezione" name="sezione">
+            <br/>
+            <button type="submit">Cerca</button>
+        </form>
+    </main>
+    <footer>
+        <p>© 2025 ProMeteo. All rights reserved.</p>
+        <button type="easter egg" id="easter-egg"></button>
+        <div id="easter-egg-popup" class="popup">
+            <a href="https://youtu.be/PbPn-Jg6u20?si=ipKQb2zNt3Eg80wJ" target="_blank">🐰</a>
+        </div>
+    </footer>
+</body>
+</html>
