+++ src/web/templates/login.html
@@ -1,14 +1,14 @@
 <html lang="it">
     <title>Login</title>
-    <link rel="stylesheet" href="loginStyle.css">
     <link href="https://unpkg.com/boxicons@2.1.4/css/boxicons.min.css" rel="stylesheet">
-    <link rel="stylesheet" href="login.js">
+    <link rel="stylesheet" href="./static/css/login.css">
+    <link rel="stylesheet" href="./static/js/login.js">
 </head>
 
 <body>
     <div class="container">
     </div class="form-box login">
-    <form action="#">
+    <form action="/login" method="POST">
         <h1>Login</h1>
         <div class="input-box">
             <input type="email" placeholder="Email" required>
