+++ src/web/templates/menu.html
@@ -1,8 +1,6 @@
 <html lang="it">
     <head>
-        <title>
-            Menu
-        </title>
+        <title>Menu</title>
         <link rel="preconnect" href="https://fonts.googleapis.com">
         <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
         <link href="https://fonts.googleapis.com/css2?family=Audiowide&display=swap" rel="stylesheet">
